Mac OSX can't convert int to size_t (which is larger other than the sign( why is openssl.h using int instead of size_t anyway)) anymore?

+++ src/netlib/ssl_layer.c
@@ -133,7 +133,7 @@ EVP_PKEY *genKey() {
 
 	{
 		OSSL_PARAM params[2] = {
-			OSSL_PARAM_BN( OSSL_PKEY_PARAM_RSA_E, bne, BN_num_bytes( bne ) ),
+			OSSL_PARAM_BN( OSSL_PKEY_PARAM_RSA_E, bne, (size_t)BN_num_bytes( bne ) ),
 			OSSL_PARAM_END
 		};
